fix Product test

+++ migrations/sqls/20220114012401-products-table-up.sql
@@ -1,5 +1,5 @@
 CREATE TABLE products (
     id SERIAL PRIMARY KEY,
     name VARCHAR(150) NOT NULL,
-    price NUMERIC(4,2) NOT NULL
+    price INTEGER NOT NULL
 );
